ransomware/requirements.txt
pycrypto
argparse
termcolor
glob2